% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse}
\alias{browse}
\title{Navigate to an ORCID profile in your default browser}
\usage{
browse(orcid)
}
\arguments{
\item{orcid}{An \code{or_cid} class object}
}
\description{
Navigate to an ORCID profile in your default browser
}
\examples{
\dontrun{
browse(as.orcid("0000-0002-1642-628X"))
}
}
