% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{extract_bibtex}
\alias{extract_bibtex}
\title{extract BibTeX record from ORCID JSON sring if available}
\usage{
extract_bibtex(x)
}
\arguments{
\item{x}{(\strong{required}): the output of \code{cite_put()}}
}
\value{
Function returns a formated BibTeX record, if nothing
was found or the BibTeX record is invalid, the input is passed
through without modifying it
}
\description{
Helper function to extract BibTeX records which may be
available in the string returned from ORCID. The function is exported.
}
\seealso{
\link[jsonlite:fromJSON]{jsonlite::fromJSON}, \code{cite_put}
}
