% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/identifiers.R
\name{identifiers}
\alias{identifiers}
\title{Get identifiers}
\usage{
identifiers(x, type = "doi", ...)
}
\arguments{
\item{x}{An object of class works, other things in the future}

\item{type}{(character) One of doi (default), pmid, pmc, other_id}

\item{...}{Ignored.}
}
\description{
Get identifiers
}
\examples{
\dontrun{
x <- works(orcid_id("0000-0001-8607-8025"))
# doi by default
identifiers(x)

# pmid
identifiers(x, "pmid")

# pmc
identifiers(x, "pmc")

# other_id
identifiers(x, "other_id")
}
}

