% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rorcid-package.R
\docType{data}
\name{issn_title}
\alias{issn_title}
\title{Lookup vector for journal titles by ISSN}
\description{
named vector of journal titles. the values are journal titles and
the names are ISSN's.
}
\details{
length: 57,968

data collected on 2018-06-13 from Crossref
}
\keyword{data}
