\name{orc}
\alias{orc}
\alias{orc.full}
\alias{orc.short}
\title{Get taxonomic hierarchy for a given taxon ID.}
\usage{
  orc(orcid, profile)

  \method{orc}{short} (orcid = NULL, profile = NULL)

  \method{orc}{full} (orcid = NULL, profile = NULL)
}
\arguments{
  \item{orcid}{}

  \item{profile}{}
}
\value{
  orcid info
}
\description{
  Get taxonomic hierarchy for a given taxon ID.
}

