% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_citations.R
\name{orcid_citations}
\alias{orcid_citations}
\title{Get citations}
\usage{
orcid_citations(orcid, put_code = NULL, ...)
}
\arguments{
\item{orcid}{(character) Orcid identifier(s) of the form
XXXX-XXXX-XXXX-XXXX. required.}

\item{put_code}{(character/integer) one or more put codes. up to
50. optional}

\item{...}{Curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\value{
A list of results for each Orcid ID passed in, with each element
named by the Orcid ID

data.frame, with the columns:
\itemize{
\item put: ORCID PUT code, identifying the work identifier in ORCID's records
\item id: the external identifier
\item id_type: the type of external identifier
\item citation: the citation as JSON
}
}
\description{
Get citations
}
\details{
This function is vectorized, so you can pass in many ORCID's, and
there's an element returned for each ORCID you put in.

This function is focused on getting citations only.
You can get all citations for an ORCID, or for certain works
using a PUT code, or for many PUT codes.

We attempt to get citations via Crossref using \pkg{rcrossref}
whenever possible as they are the most flexible and don't have as
many mistakes in the text. If there is no DOI, we fetch the
citation from ORCID.

Right now we get JSON citations back. We'd like to support bibtex
format. DOI.org supports this but not ORCID.
}
\examples{
\dontrun{
(res <- orcid_citations(orcid = "0000-0002-9341-7985"))
(res2 <- orcid_citations(orcid = "0000-0002-1642-628X"))
(res2 <- orcid_citations(orcid = c("0000-0002-9341-7985", "0000-0002-1642-628X")))

# get individual works
## a single put code
(a <- orcid_citations(orcid = "0000-0002-9341-7985", put_code = 5011717))
## many put codes
(b <- orcid_citations(orcid = "0000-0002-9341-7985", 
   put_code = c(5011717, 15536016)))
}
}
