% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_ping.R
\name{orcid_ping}
\alias{orcid_ping}
\title{Check if ORCID API is up and running}
\usage{
orcid_ping(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[crul:HttpClient]{crul::HttpClient()}}}
}
\value{
a text string
}
\description{
Check if ORCID API is up and running
}
\examples{
\dontrun{
orcid_ping()
}
}
