% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rorcid-package.R
\name{rorcid-defunct}
\alias{rorcid-defunct}
\title{Defunct functions in rorcid}
\description{
\itemize{
 \item \code{\link{summary.or_cid}}: Function is gone. Deemed not really 
 that useful, and hard to maintain given other changes in the package.
}
}
