% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{rorcid-package}
\alias{rorcid-package}
\title{A programmatic R interface the Orcid.org API.}
\description{
A programmatic R interface the Orcid.org API.
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

