% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/works.R
\name{works}
\alias{works}
\title{Get works data}
\usage{
works(x)
}
\arguments{
\item{x}{Input from a call to \code{\link{orcid_id}} or \code{\link{as.orcid}}}

\item{...}{Ignored.}
}
\description{
From a call to orcid_id or as.orcid
}
\examples{
\dontrun{
out <- works(orcid_id("0000-0002-9341-7985"))
out
head(out$data)
works( orcid_id("0000-0003-1620-1408") )
works( orcid_id("0000-0002-1642-628X") )
works( orcid_id("0000-0003-1444-9135") )
works( orcid_id("0000-0003-1419-2405") )

out <- orcid(query="keyword:ecology")
works(orcid_id(out$data$`orcid-identifier.path`[10]))

works(as.orcid("0000-0002-1642-628X"))
}
}

